Add a 'keep networks' bulldoze mode
This is an alternate mode for the network bulldoze tool, it is
activated by pressing the Shift key when releasing the left mouse
button after selecting an area.
When in this mode, the bulldoze tool will remove everything other
than the transportation networks.

If the Shift key is not pressed when releasing the left mouse button
after selecting an area, the network bulldoze tool operates in its
existing 'remove only transportation networks' mode.